fix: 자동차 이름 다섯 글자 제한
