+++ src/main/java/racing/control/RaceOperator.java
@@ -5,13 +5,16 @@ import racing.domain.RaceResult;
 import racing.view.InputView;
 import racing.view.OutputView;
 
+import java.util.List;
+import java.util.stream.Collectors;
+
 public class RaceOperator {
     private static Race race;
     private int numTrials;
 
     public void setupGame() {
         race = new Race();
-        race.createCars();
+        setupCar();
         numTrials = inputNumTrials();
     }
 
@@ -24,6 +27,32 @@ public class RaceOperator {
             System.out.println(e.getMessage());
             return inputNumTrials();
         }
+    }
+
+    public void setupCar() {
+        try {
+            race.createCars(inputCarNames());
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            setupCar();
+        }
+    }
+
+    private List<String> inputCarNames() {
+        try {
+            List<String> carNames = InputView.requestNames();
+            checkDuplicatedNames(carNames);
+            return carNames;
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            return inputCarNames();
+        }
+    }
+
+    private boolean checkDuplicatedNames(final List<String> names) {
+        List<String> reducedNames = names.stream().distinct().collect(Collectors.toList());
+        if (names.size() != reducedNames.size()) throw new IllegalArgumentException("중복된 이름이 존재하면 안됩니다.");
+        return true;
     }
 
     public void startGame() {
